Rename well behaved plugin test class to better represent their scenario
