+++ subprojects/language-base/src/functionalTest/groovy/dev/nokee/language/base/LanguageBasePluginWellBehavedPluginTest.java
@@ -23,7 +23,7 @@ import org.junit.jupiter.api.TestFactory;
 
 import java.util.stream.Stream;
 
-class LanguageBasePluginFunctionalTest {
+class LanguageBasePluginWellBehavedPluginTest {
 	@TestFactory
 	Stream<DynamicTest> checkWellBehavedPlugin() {
 		return new WellBehavedPluginTester()
