+++ subprojects/language-cpp/src/functionalTest/groovy/dev/nokee/language/cpp/CppLanguagePluginWellBehavedPluginTest.java
@@ -13,21 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.language.objectivec;
+package dev.nokee.language.cpp;
 
 import dev.nokee.internal.testing.WellBehavedPluginTester;
 import dev.nokee.internal.testing.util.TestCaseUtils;
-import dev.nokee.language.objectivec.internal.plugins.ObjectiveCLanguageBasePlugin;
+import dev.nokee.language.cpp.internal.plugins.CppLanguagePlugin;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 
 import java.util.stream.Stream;
 
-class ObjectiveCLanguageBasePluginFunctionalTest {
+class CppLanguagePluginWellBehavedPluginTest {
 	@TestFactory
 	Stream<DynamicTest> checkWellBehavedPlugin() {
 		return new WellBehavedPluginTester()
-			.pluginClass(ObjectiveCLanguageBasePlugin.class)
+			.qualifiedPluginId("dev.nokee.cpp-language")
+			.pluginClass(CppLanguagePlugin.class)
 			.stream().map(TestCaseUtils::toJUnit5DynamicTest);
 	}
 }
