+++ subprojects/language-objective-cpp/src/functionalTest/groovy/dev/nokee/language/objectivecpp/ObjectiveCppLanguageBasePluginWellBehavedPluginTest.java
@@ -13,21 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.model;
+package dev.nokee.language.objectivecpp;
 
 import dev.nokee.internal.testing.WellBehavedPluginTester;
 import dev.nokee.internal.testing.util.TestCaseUtils;
-import dev.nokee.model.internal.plugins.ModelBasePlugin;
+import dev.nokee.language.objectivecpp.internal.plugins.ObjectiveCppLanguageBasePlugin;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 
 import java.util.stream.Stream;
 
-class ModelBasePluginFunctionalTest {
+class ObjectiveCppLanguageBasePluginWellBehavedPluginTest {
 	@TestFactory
 	Stream<DynamicTest> checkWellBehavedPlugin() {
 		return new WellBehavedPluginTester()
-			.pluginClass(ModelBasePlugin.class)
+			.pluginClass(ObjectiveCppLanguageBasePlugin.class)
 			.stream().map(TestCaseUtils::toJUnit5DynamicTest);
 	}
 }
