+++ subprojects/language-objective-cpp/src/functionalTest/groovy/dev/nokee/language/objectivecpp/ObjectiveCppLanguagePluginWellBehavedPluginTest.java
@@ -13,22 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.runtime.nativebase;
+package dev.nokee.language.objectivecpp;
 
 import dev.nokee.internal.testing.WellBehavedPluginTester;
 import dev.nokee.internal.testing.util.TestCaseUtils;
-import dev.nokee.runtime.nativebase.internal.NativeRuntimePlugin;
+import dev.nokee.language.objectivecpp.internal.plugins.ObjectiveCppLanguagePlugin;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 
 import java.util.stream.Stream;
 
-class NativeRuntimePluginFunctionalTest {
+class ObjectiveCppLanguagePluginWellBehavedPluginTest {
 	@TestFactory
 	Stream<DynamicTest> checkWellBehavedPlugin() {
 		return new WellBehavedPluginTester()
-			.qualifiedPluginId("dev.nokee.native-runtime")
-			.pluginClass(NativeRuntimePlugin.class)
+			.qualifiedPluginId("dev.nokee.objective-cpp-language")
+			.pluginClass(ObjectiveCppLanguagePlugin.class)
 			.stream().map(TestCaseUtils::toJUnit5DynamicTest);
 	}
 }
