+++ subprojects/platform-jni/src/functionalTest/groovy/dev/nokee/platform/jni/JniLibraryPluginWellBehavedPluginTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 the original author or authors.
+ * Copyright 2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.language.c;
+package dev.nokee.platform.jni;
 
 import dev.nokee.internal.testing.WellBehavedPluginTester;
 import dev.nokee.internal.testing.util.TestCaseUtils;
-import dev.nokee.language.c.internal.plugins.CLanguageBasePlugin;
+import dev.nokee.platform.jni.internal.plugins.JniLibraryPlugin;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 
 import java.util.stream.Stream;
 
-class CLanguageBasePluginFunctionalTest {
+class JniLibraryPluginWellBehavedPluginTest {
 	@TestFactory
 	Stream<DynamicTest> checkWellBehavedPlugin() {
 		return new WellBehavedPluginTester()
-			.pluginClass(CLanguageBasePlugin.class)
+			.qualifiedPluginId("dev.nokee.jni-library")
+			.pluginClass(JniLibraryPlugin.class)
 			.stream().map(TestCaseUtils::toJUnit5DynamicTest);
 	}
 }
