+++ subprojects/runtime-base/src/functionalTest/groovy/dev/nokee/runtime/base/RuntimeBasePluginWellBehavedPluginTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,22 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.platform.jni;
+package dev.nokee.runtime.base;
 
 import dev.nokee.internal.testing.WellBehavedPluginTester;
 import dev.nokee.internal.testing.util.TestCaseUtils;
-import dev.nokee.platform.jni.internal.plugins.JniLibraryPlugin;
+import dev.nokee.runtime.base.internal.RuntimeBasePlugin;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 
 import java.util.stream.Stream;
 
-class JniLibraryWellBehavedPluginTest {
+class RuntimeBasePluginWellBehavedPluginTest {
 	@TestFactory
 	Stream<DynamicTest> checkWellBehavedPlugin() {
 		return new WellBehavedPluginTester()
-			.qualifiedPluginId("dev.nokee.jni-library")
-			.pluginClass(JniLibraryPlugin.class)
+			.pluginClass(RuntimeBasePlugin.class)
 			.stream().map(TestCaseUtils::toJUnit5DynamicTest);
 	}
 }
