+++ subprojects/runtime-darwin/src/functionalTest/groovy/dev/nokee/runtime/darwin/DarwinRuntimePluginWellBehavedPluginTest.java
@@ -13,21 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.language.swift;
+package dev.nokee.runtime.darwin;
 
 import dev.nokee.internal.testing.WellBehavedPluginTester;
 import dev.nokee.internal.testing.util.TestCaseUtils;
-import dev.nokee.language.swift.internal.plugins.SwiftLanguageBasePlugin;
+import dev.nokee.runtime.darwin.internal.DarwinRuntimePlugin;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 
 import java.util.stream.Stream;
 
-class SwiftLanguageBasePluginFunctionalTest {
+class DarwinRuntimePluginWellBehavedPluginTest {
 	@TestFactory
 	Stream<DynamicTest> checkWellBehavedPlugin() {
 		return new WellBehavedPluginTester()
-			.pluginClass(SwiftLanguageBasePlugin.class)
+			.qualifiedPluginId("dev.nokee.darwin-runtime")
+			.pluginClass(DarwinRuntimePlugin.class)
 			.stream().map(TestCaseUtils::toJUnit5DynamicTest);
 	}
 }
