+++ subprojects/runtime-native/src/functionalTest/groovy/dev/nokee/runtime/nativebase/NativeRuntimeBasePluginWellBehavedPluginTest.java
@@ -13,22 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.language.objectivec;
+package dev.nokee.runtime.nativebase;
 
 import dev.nokee.internal.testing.WellBehavedPluginTester;
 import dev.nokee.internal.testing.util.TestCaseUtils;
-import dev.nokee.language.objectivec.internal.plugins.ObjectiveCLanguagePlugin;
+import dev.nokee.runtime.nativebase.internal.NativeRuntimeBasePlugin;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 
 import java.util.stream.Stream;
 
-class ObjectiveCLanguagePluginFunctionalTest {
+class NativeRuntimeBasePluginWellBehavedPluginTest {
 	@TestFactory
 	Stream<DynamicTest> checkWellBehavedPlugin() {
 		return new WellBehavedPluginTester()
-			.qualifiedPluginId("dev.nokee.objective-c-language")
-			.pluginClass(ObjectiveCLanguagePlugin.class)
+			.qualifiedPluginId("dev.nokee.native-runtime-base")
+			.pluginClass(NativeRuntimeBasePlugin.class)
 			.stream().map(TestCaseUtils::toJUnit5DynamicTest);
 	}
 }
